+++ src/main/java/com/example/orderparser/batchprocessor/ConsoleWriter.java
@@ -1,7 +1,6 @@
-package com.example.orders_parser;
+package com.example.orderparser.batchprocessor;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import com.example.orderparser.Order;
 import org.springframework.batch.item.ItemWriter;
 import org.springframework.stereotype.Component;
 
@@ -9,11 +8,9 @@ import java.util.List;
 
 @Component
 public class ConsoleWriter implements ItemWriter<Order> {
-    private static final Logger log = LoggerFactory.getLogger(ConsoleWriter.class);
 
     @Override
-    public void write(List<? extends Order> list) throws Exception {
-        log.info("Writing");
+    public void write(List<? extends Order> list) {
         for (Order order : list) {
             System.out.println(order.toString());
         }
