+++ src/main/java/com/example/orderparser/configuration/OrderProcessorConfiguration.java
@@ -1,6 +1,9 @@
-package com.example.orders_parser.configuration;
+package com.example.orderparser.configuration;
 
-import com.example.orders_parser.*;
+import com.example.orderparser.*;
+import com.example.orderparser.batchprocessor.ConsoleWriter;
+import com.example.orderparser.batchprocessor.LineNumberMapper;
+import com.example.orderparser.batchprocessor.LineProcessor;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.Step;
 import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
@@ -14,7 +17,6 @@ import org.springframework.batch.item.ItemWriter;
 import org.springframework.batch.item.file.FlatFileItemReader;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
