--- src/main/java/com/example/orders_parser/OrderParser.java
@@ -1,7 +1,0 @@
-package com.example.orders_parser;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-
-public interface OrderParser {
-    Order parse(String line, Extensions ext) throws JsonProcessingException;
-}
