+++ src/main/resources/application.properties
@@ -1,11 +1,1 @@
 maxThreads = 2
-
-logging.level.org.springframework.web=ERROR
-logging.file.name=${java.io.tmpdir}/order_parser.log
-
-spring.datasource.url = jdbc:mysql://localhost/abc
-spring.datasource.name=testme
-spring.datasource.username=xxxx
-spring.datasource.password=xxxx
-spring.datasource.driver-class-name= com.mysql.jdbc.Driver spring.jpa.database=mysql
-spring.jpa.database-platform=org.hibernate.dialect.MySQLDialect
