+++ src/test/resources/expected_csv.txt
@@ -1,4 +1,4 @@
-{"id":1,"amount":100.0,"comment":"оплата заказаcsv2","filename":"sample.csv","line":2,"result":"OK"}
+{"id":1,"amount":100.0,"comment":"оплата заказаcsv2","filename":"sample.csv","line":1,"result":"OK"}
 {"id":7,"amount":1020.0,"comment":"оплата заказаcsv3","filename":"sample.csv","line":2,"result":"OK"}
 {"id":6,"amount":103230.0,"comment":"оплата заказаcsv4","filename":"sample.csv","line":3,"result":"OK"}
 {"id":2,"amount":500.0,"comment":"оплата заказаcsv5","filename":"sample.csv","line":4,"result":"OK"}
