--- src/test/resources/logback-test.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-    <include resource="org/springframework/boot/logging/logback/base.xml" />
-    <logger name="org.springframework" level="OFF"/>
-    <root level="OFF">
-        <appender-ref ref="STDOUT"/>
-    </root>
-</configuration>
